% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{project-package}
\alias{project}
\alias{project-package}
\title{project: Work with European Football Data}
\description{
Standardize and merge data from european football clubs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/samtaukecornerstone/project}
  \item Report bugs at \url{https://github.com/samtaukecornerstone/project/issues}
}

}
\author{
\strong{Maintainer}: Sam Tauke \email{stauke@cornerstone.com}

}
\keyword{internal}
