% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-matches.R
\name{uss_countries}
\alias{uss_countries}
\title{Get available countries}
\usage{
uss_countries()
}
\value{
\code{character} vector containing names of available countries.
}
\description{
These are the countries for which there is league-play data available
from {engsoccerdata}.
}
\examples{
uss_countries()
}
