% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{uss_get_games}
\alias{uss_get_games}
\title{Get a league-play tibble}
\usage{
uss_get_games(country = uss_countries())
}
\arguments{
\item{country}{\code{character} scalar, used to populate \code{country} column}
}
\value{
\code{tbl_df} with columns \code{country}, \code{date}, \code{season}, \code{tier},
\code{home}, \code{visitor}, \code{goals_home}, \code{goals_visitor}.
}
\description{
Get a league-play tibble
}
\examples{
uss_get_games("england")
}
