% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{uss_make_games}
\alias{uss_make_games}
\title{Make a standard league-play tibble}
\usage{
uss_make_games(data_engsoc, country)
}
\arguments{
\item{data_engsoc}{\code{data.frame} obtained from {engsoccerdata}.}

\item{country}{\code{character} scalar, used to populate \code{country} column}
}
\value{
\code{tbl_df} with columns \code{country}, \code{date}, \code{season}, \code{tier},
\code{home}, \code{visitor}, \code{goals_home}, \code{goals_visitor}.
}
\description{
Make a standard league-play tibble
}
\examples{
uss_make_games(engsoccerdata::italy, "Italy")
}
\keyword{internal}
