% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{uss_make_matches}
\alias{uss_make_matches}
\title{Make a standard league-play tibble}
\usage{
uss_make_matches(data_engsoc, country)
}
\arguments{
\item{data_engsoc}{obtained from {engsoccerdata}.}

\item{country}{`character` scalar, specifies the league.}
}
\value{
a tibble with columns `country`, `date`, `season`, `tier`, `home`,
   `visitor`, `goals_home`, `goals_visitor`.
}
\description{
Given a league-play data frame from {engsoccer}, returns a tibble with
standardised column-names and types, e.g. `date` is a `Date`.
}
\examples{
uss_make_matches(engsoccerdata::spain,"Spain")
}
