% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{uss_make_seasons_cumulative}
\alias{uss_make_seasons_cumulative}
\alias{uss_make_seasons_final}
\title{Make season-based calculations}
\usage{
uss_make_seasons_cumulative(
  data_teams_matches,
  fn_points_per_win = uss_points_per_win
)

uss_make_seasons_final(
  data_teams_matches,
  fn_points_per_win = uss_points_per_win
)
}
\arguments{
\item{data_teams_matches}{data frame created using \code{\link[=uss_make_teams_matches]{uss_make_teams_matches()}}}

\item{fn_points_per_win}{\code{function} with vectorized arguments \code{country},
\code{season}, that returns a integer indicating points-per-win.
A default function is provided, \code{\link[=uss_points_per_win]{uss_points_per_win()}}, which includes
the countries in \code{\link[=uss_countries]{uss_countries()}}.}
}
\value{
\link[tibble:tibble-package]{tibble} with columns
\code{country}, \code{tier}, \code{season}, \code{team}, \code{date}, \code{matches}, \code{wins},
\code{draws}, \code{losses}, \code{points}, \code{goals_for}, \code{goals_against}.
}
\description{
Given a teams-matches data frame (returned by \code{\link[=uss_make_teams_matches]{uss_make_teams_matches()}}),
return return a data frame on wins, losses, points, etc.:
\itemize{
\item cumulative, over the course of each season: \code{uss_make_seasons_cumulative()}
}
}
\examples{
italy <- uss_get_matches("italy") |> uss_make_teams_matches()
uss_make_seasons_cumulative(italy)
uss_make_seasons_final(italy)
}
