% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams-matches.R
\name{uss_make_teams_matches}
\alias{uss_make_teams_matches}
\title{Make a teams-matches tibble}
\usage{
uss_make_teams_matches(data_matches)
}
\arguments{
\item{data_matches}{\code{data.frame} created using \code{uss_get_matches()} or \code{uss_make_matches()}}
}
\value{
\link[tibble:tibble-package]{tibble} with columns \code{country}, \code{tier},
\code{season}, \code{team}, \code{date}, \code{opponent}, \code{at_home},
\code{goals_for}, \code{goals_against}; ordered by \code{country}, \code{tier},
\code{season}, \code{team}, \code{date}.
}
\description{
This has essentially the same as the \code{matches} data; here, a row is returned
for every combination of team and match. Thus there are two rows returned
for each row in the \code{data_matches} tibble: one for each team.
}
\examples{
uss_get_matches("england") |>
  uss_make_teams_matches() |>
  dplyr::filter(team == "Leeds United") |>
  tail()
}
