% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{uss_plot_seasons_tiers}
\alias{uss_plot_seasons_tiers}
\title{Plot over season-final results}
\usage{
uss_plot_seasons_tiers(data_seasons, aes_y = .data$wins, ncol = 1)
}
\arguments{
\item{data_seasons, }{data frame created using \code{\link[=uss_make_seasons_final]{uss_make_seasons_final()}}.}

\item{aes_y, }{\verb{<data-masking>} expression used for the y-aesthetic.}

\item{ncol, }{\code{integer}-ish number of columns in facet.}
}
\value{
Object with S3 classes \code{"gg"}, \code{"ggplot"}, i.e. a ggplot2 object.
}
\description{
Returns a ggplot:
\itemize{
\item \code{season} on the x-axis
\item faceted by \code{team}
\item has two layers:
\itemize{
\item rectangles shaded by tier
\item points showing some measure (default \code{wins}) on the y-axis
}
}
}
\examples{
seas <-
  uss_get_matches("england") |>
  uss_make_teams_matches() |>
  dplyr::filter(team \%in\% c("Leeds United", "Norwich City")) |>
  uss_make_seasons_final() |>
  dplyr::arrange(team, season)

# use default (wins)
uss_plot_seasons_tiers(seas)

# use custom expression
uss_plot_seasons_tiers(seas, goals_for - goals_against)
}
