% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{uss_points}
\alias{uss_points}
\alias{uss_points_per_win}
\title{Calculate points}
\usage{
uss_points(wins, draws, country = NULL, season = NULL)

uss_points_per_win(country, season)
}
\arguments{
\item{wins}{\code{integer} number of wins}

\item{draws}{\code{integer} number of draws}

\item{country}{\code{character} scalar, specifies the league.
\code{uss_countries()} returns choices available from {engsoccerdata}}

\item{season}{\code{integer} year of the football season.}
}
\value{
\code{integer}:
\describe{
\item{\code{uss_points()}}{number of points}
\item{\code{uss_points_per_win()}}{number of points per win}
}
}
\description{
Calculate points
}
\examples{
uss_points_per_win("england", 1980)
uss_points(3, 1, country = "england", season = 1980)
}
