% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{uss_points_per_win}
\alias{uss_points_per_win}
\title{Calculate points}
\usage{
uss_points_per_win(country, season)
}
\arguments{
\item{country}{\code{character} scalar, specifies the league.
\code{uss_countries()} returns choices available from {engsoccerdata}}

\item{season}{\code{integer}-ish year of the football season.}
}
\value{
\code{integer} number of points per win
}
\description{
Calculate points
}
\examples{
uss_points_per_win("england", 1980)
uss_points_per_win(c("england", "england"), c(1980, 1981))
}
