% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{uss_result}
\alias{uss_result}
\title{Structure for printing results}
\usage{
uss_result(goals_for = integer(), goals_against = integer())
}
\arguments{
\item{goals_for}{\code{integer}-ish number of goals for a team in a match.}

\item{goals_against}{\code{integer}-ish number of goals against a team in a match.}
}
\value{
An S3 object with class \code{ussie_result}.
}
\description{
For \code{teams_matches} tibbles, it can be convenient to see the result as a
structure, rather than using only \code{goals_for} and \code{goals_against}.
}
\details{
This function is vectorized over \code{goals_for} and \code{goals_against}; they must
have the same length.
}
\examples{
uss_result(3, 2)
uss_get_matches("italy") |>
  uss_make_teams_matches() |>
  dplyr::mutate(
    result = uss_result(goals_for, goals_against),
    .after = opponent
  )
}
