% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Working with European Football Data}
\description{
This is a package that lets you play with European football data. It is for a workshop and more or less an identical copy of the Building Tidy Tools workshop demo package from rstudio::conf2022. This copy is incredibly boring and there is no reason you would use it. If you want to follow the workshop, do that and ignore my personal repo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kpisane/ussie}
  \item \url{https://kpisane.github.io/ussie/}
  \item Report bugs at \url{https://github.com/kpisane/ussie/issues}
}

}
\author{
\strong{Maintainer}: Kelly Pisane \email{kellypisane@outlook.com}

}
\keyword{internal}
