% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie2-package.R
\docType{package}
\name{ussie2-package}
\alias{ussie2}
\alias{ussie2-package}
\title{ussie2: Tools for publicly available football data}
\description{
This is the second version of the `ussie` package that was demo'd at the Building Tidy Tools workshop at #rstudioconf2022.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jpflores-13/ussie2}
  \item Report bugs at \url{https://github.com/jpflores-13/ussie2/issues}
}

}
\author{
\strong{Maintainer}: JP Flores \email{jpflores013@gmail.com} (\href{https://orcid.org/0000-0001-5619-8990}{ORCID})

}
\keyword{internal}
