% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-matches.R
\name{uss_get_matches}
\alias{uss_get_matches}
\title{Get a league-play tibble}
\usage{
uss_get_matches(country = uss_countries())
}
\arguments{
\item{country}{\code{character} scalar, specifies the league.
\code{uss_countries()} returns choices available from {engsoccerdata}}
}
\value{
\link[tibble:tibble-package]{tibble} with columns \code{country}, \code{date},
\code{season}, \code{tier}, \code{home}, \code{visitor}, \code{goals_home}, \code{goals_visitor}.
}
\description{
Gets league-play data for each game, from {engsoccerdata}, returning
a tibble in a standardised format.
}
\details{
\code{uss_countries()} returns the available choices; \code{"england"} is
the default.

This function relies on an internal function, \code{uss_make_matches()}, to parse
the source data.
}
\examples{
uss_get_matches("england")
}
