% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{uss_get_matches}
\alias{uss_get_matches}
\title{Get a league-play tibble}
\usage{
uss_get_matches(country = uss_countries(), ...)
}
\arguments{
\item{country}{\code{character} scalar, specifies the league.
\code{uss_countries()} returns choices available from {engsoccerdata}}

\item{...}{
  Arguments passed on to \code{\link[dplyr:filter]{dplyr::filter}}
  \describe{
    \item{\code{.data}}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}
    \item{\code{.preserve}}{Relevant when the \code{.data} input is grouped.
If \code{.preserve = FALSE} (the default), the grouping structure
is recalculated based on the resulting data, otherwise the grouping is kept as is.}
  }}
}
\value{
\link[tibble:tibble-package]{tibble} with columns \code{country}, \code{date},
}
\description{
Gets league-play data for each game, from {engsoccerdata}, returning
a tibble in a standardised format. You can pass in filtering expressions
via the dots (\code{...}), these are evaluated using \code{\link[dplyr:filter]{dplyr::filter()}}.
}
\details{
\code{uss_countries()} returns the available choices; \code{"england"} is
the default.

This function relies on an internal function, \code{uss_make_matches()}, to parse
the source data; to speed up the performance, a memoised (caching) version
of this function is used, \code{uss_make_matches_mem()}.
}
\examples{
uss_get_matches("england")
}
