% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{uss_make_matches}
\alias{uss_make_matches}
\title{Title Make a standard league-play tibble}
\usage{
uss_make_matches(data_engsoc)
}
\arguments{
\item{data_engsoc}{df from engsoccerdata package}

\item{country}{a country one of spain, uk, Italy, Germany, holland}
}
\value{
a tibble with columns \code{country}, \code{date}, \code{season}, \code{tier}, \code{home},
\code{visitor}, \code{goals_home}, \code{goals_visitor}.
}
\description{
Title Make a standard league-play tibble
}
\examples{
uss_make_matches(engsoccerdata::spain, "Spain")
}
