% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{uss_make_matches}
\alias{uss_make_matches}
\title{Create a standard dataframe for league data}
\usage{
uss_make_matches(data_engsoc, country_name)
}
\arguments{
\item{data_engsoc}{Data produced by engsoccerdata}

\item{country_name}{Name of country that will be added to output to
identify country used}
}
\value{
a tibble with columns \code{country}, \code{date}, \code{season}, \code{tier}, \code{home},
\code{visitor}, \code{goals_home}, \code{goals_visitor}.
}
\description{
Create a standard dataframe for league data
}
