% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{uss_make_matches}
\alias{uss_make_matches}
\title{Makes standard dataframe for league data}
\usage{
uss_make_matches(data_engsoc, country)
}
\arguments{
\item{data_engsoc}{dataframe from engsoccer data package}

\item{country}{a country one of spain, italy, germany, england, holland}
}
\value{
dataframe, tibble
}
\description{
Makes standard dataframe for league data
}
