% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{uss_make_matches}
\alias{uss_make_matches}
\title{Format European Football data to standard format}
\usage{
uss_make_matches(data_engsoc, country)
}
\arguments{
\item{data_engsoc}{\verb{data frame} from {engsoccerdata}.}

\item{country}{\code{character} country name, one of Spain,  Italy, Germany, England, Holland}
}
\value{
dataframe, tibble with columns \code{country},
\code{date}, \code{season}, \code{tier}, \code{home}, \code{visitor},
\code{goals_home}, \code{goals_visitor}.
}
\description{
Format European Football data to standard format
}
\examples{
requireNamespace("engsoccerdata", quietly = TRUE)
uss_make_matches(engsoccerdata::england, "England")
}
