% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{uss_make_matches}
\alias{uss_make_matches}
\title{Makes a standard dataframe for league data
Please note holland is not a country, I will fix that later, this is a quirk
of the source data}
\usage{
uss_make_matches(data_engsoc, country)
}
\arguments{
\item{data_engsoc}{dataframe from engsoccerdata package}

\item{country}{a country (one of spain, italy, germany, holland, england)}
}
\value{
a tibble with columns \code{country}, \code{date}, \code{season}, \code{tier}, \code{home},
\code{visitor}, \code{goals_home}, \code{goals_visitor}.
}
\description{
Makes a standard dataframe for league data
Please note holland is not a country, I will fix that later, this is a quirk
of the source data
}
\examples{
uss_make_matches(engsoccerdata::spain, "Spain")
}
