% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{uss_make_matches}
\alias{uss_make_matches}
\title{Make a standard league-play tibble}
\usage{
uss_make_matches(data_engsoc, country)
}
\arguments{
\item{data_engsoc}{data frame from engsoccerdata package}

\item{country}{Spain, Italy, Germany, England & Holland}
}
\value{
a tibble with columns: \code{country}, \code{date}, \code{season}, \code{tier}, \code{home}, \code{visitor}, \code{goals_home}, \code{goals_visitor}.
}
\description{
Make a standard league-play tibble
}
\examples{
uss_make_matches(engsoccerdata::spain, "Spain")
}
