% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{uss_make_matches}
\alias{uss_make_matches}
\title{Makes a df for league data}
\usage{
uss_make_matches(data_engsoc, country)
}
\arguments{
\item{data_engsoc}{dataframe from engsoccerdata package}

\item{country}{country \code{character} scalar, specifies the league.}
}
\value{
dataframe, a tibble with columns \code{country}, \code{date}, \code{season}, \code{tier}, \code{home},
\code{visitor}, \code{goals_home}, \code{goals_visitor}.
}
\description{
Given a league-play data frame from {engsoccer}, returns a tibble with
standardised column-names and types, e.g. \code{date} is a \code{Date}.
}
