% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{uss_make_matches}
\alias{uss_make_matches}
\title{Makes a standard dataframe for league data}
\usage{
uss_make_matches(data_engsoc, country)
}
\arguments{
\item{data_engsoc}{dataframe from engsoccerdata package}

\item{country}{country \code{character} specifies the league.
one of spain, italy, germany, england, holland}
}
\value{
a tibble with columns \code{country}, \code{date}, \code{season}, \code{tier}, \code{home},
\code{visitor}, \code{goals_home}, \code{goals_visitor}.
}
\description{
Makes a standard dataframe for league data
}
\examples{
uss_make_matches(engsoccerdata::england, "England")
}
