% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{uss_make_matches}
\alias{uss_make_matches}
\title{Create a standardized tibble for league data}
\usage{
uss_make_matches(data_engsoc, country)
}
\arguments{
\item{data_engsoc}{dataframe from engsoccer package}

\item{country}{league country name, one of Spain, Italy, England,
or Holland}
}
\value{
dataframe, tibble
}
\description{
Given a league-play data frame from {engsoccer}, returns a tibble with
standardised column-names and types, e.g. \code{date} is a \code{Date}.
}
\examples{
uss_make_matches(engsoccerdata::spain, "Spain")
}
