% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{uss_make_matches}
\alias{uss_make_matches}
\title{Makes a standard dataframe for league data}
\usage{
uss_make_matches(data_engsoc, country)
}
\arguments{
\item{data_engsoc}{dataframe from {engsoccerdata} package}

\item{country}{country (Spain, Italy, Germany, Holland, England)}
}
\value{
a tibble with columns \code{country}, \code{date}, \code{season}, \code{tier}, \code{home},
\code{visitor}, \code{goals_home}, \code{goals_visitor}.
}
\description{
Makes a standard dataframe for league data
}
