% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{uss_points_per_win}
\alias{uss_points_per_win}
\title{Calculate points}
\usage{
uss_points_per_win(country, season, default = 3)
}
\arguments{
\item{country}{\code{character} scalar, specifies the league.
\code{uss_countries()} returns choices available from {engsoccerdata}}

\item{season}{\code{integer}-ish year of the football season.}

\item{default}{\code{integer}-ish default value for points-per-win, used if
a value of \code{country} is not in \code{uss_countries()}.}
}
\value{
\code{integer} number of points per win
}
\description{
Given \code{country} and \code{season}, return the number of points for a win.
}
\details{
This function is vectorized over \code{country} and \code{season}; they should have
the same length. If any \code{country} is not in \code{uss_countries}, a warning
is issued, and the \code{default} value is used.
}
\examples{
uss_points_per_win("england", 1980)
uss_points_per_win(c("england", "england"), c(1980, 1981))
}
