% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Work with European Football League Data}
\description{
This is a demo package for the "Building Tidy Tools" workshop at at rstudio::conf(2022L). It allows you to work with European football league data supplied by the engsoccerdata package (Curley 2016). I like lunch.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/3mmaRand/ussie}
  \item \url{https://3mmarand.github.io/ussie/}
  \item Report bugs at \url{https://github.com/3mmaRand/ussie/issues}
}

}
\author{
\strong{Maintainer}: Emma Rand \email{emma.rand@york.ac.uk} (\href{https://orcid.org/0000-0001-6452-2420}{ORCID})

}
\keyword{internal}
