% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Work with European Soccer Data}
\description{
This is a toy package developed at rstudio::conf 2022's Building Tidy Tools workshop. It allows you to work with European football ?uss_make_matchesleague data supplied by the engsoccerdata package (Curley 2016).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ali-day/ussie}
  \item Report bugs at \url{https://github.com/ali-day/ussie/issues}
}

}
\author{
\strong{Maintainer}: Ali Day \email{aliyuenday@gmail.com}

}
\keyword{internal}
