% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Stanardizing and Transforming European Football Data}
\description{
Here is a sample package created at the 2022 Tidy Tools Workshop in Washington, DC! We will standardize and transform European football (soccer) data to compare across messy datasets. This is a way to practice using the package development tooling in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zoews/ussie}
  \item \url{https://zoews.github.io/ussie/}
  \item Report bugs at \url{https://github.com/zoews/ussie/issues}
}

}
\author{
\strong{Maintainer}: Cass Wilkinson Saldana \email{cass.z.ws@gmail.com}

}
\keyword{internal}
