% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Explore European Football Data}
\description{
Playing around with european futbol data to exemplify package development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cpenaloza/ussie}
  \item Report bugs at \url{https://github.com/cpenaloza/ussie/issues}
}

}
\author{
\strong{Maintainer}: Claudia Penaloza \email{claudiapenaloza@gmail.com}

}
\keyword{internal}
