% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Analysing Footbal Data From Europe}
\description{
Data preparation for football data from European countries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edward-burn/ussie}
  \item Report bugs at \url{https://github.com/edward-burn/ussie/issues}
}

}
\author{
\strong{Maintainer}: Edward Burn \email{edward.burn@ndorms.ox.ac.uk}

}
\keyword{internal}
