% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: working with football data}
\description{
This is a package that sort of works behind the VPN.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jamann426/ussie}
  \item Report bugs at \url{https://github.com/jamann426/ussie/issues}
}

}
\author{
\strong{Maintainer}: Jordan Amann \email{jordan.amann@fema.dhs.gov}

}
\keyword{internal}
