% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Interface for European Football League Data - rstudioconf}
\description{
Demo package from rstudio::conf(2022) workshop Building Tidy Tools. Standardizes data from the engsoccerdata package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KatherineCox/ussie}
  \item Report bugs at \url{https://github.com/KatherineCox/ussie/issues}
}

}
\author{
\strong{Maintainer}: Katherine Cox \email{katie.el.cox@gmail.com}

}
\keyword{internal}
