% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Working with European Football Data}
\description{
This is a demo package that was made as part of the 2022 RStudio Conference, during the Building Tidy Tools workshop. This package is meant to clean and analyze data from the engsoccerdata package (Curley 2016).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kathy-nevola/ussie}
  \item \url{https://kathy-nevola.github.io/ussie/}
  \item Report bugs at \url{https://github.com/kathy-nevola/ussie/issues}
}

}
\author{
\strong{Maintainer}: Kathy Nevola \email{kathy.nevola@olink.com}

}
\keyword{internal}
