% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Working with European Football Data}
\description{
Built as a part of rstudio::conf2022 "Building Tidy Tools" workshop. Package originally built by Ian Lyttle and gone over by Emma Rand.
}
\author{
\strong{Maintainer}: Marina Lehmkuhl \email{dustin.lehmkuhl@regeneron.com}

}
\keyword{internal}
