% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Tidies Football data}
\description{
This is a demo package for the "Building Tidy Tools" workshop at at rstudio::conf(2022L). It allows you to work with European football league data supplied by the engsoccerdata package (Curley 2016).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maracena/ussie}
  \item Report bugs at \url{https://github.com/maracena/ussie/issues}
}

}
\author{
\strong{Maintainer}: Miguel  Aracena Velazquez \email{maaracena27@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
