% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: European Football Data Manipulation}
\description{
Manipulation of Football/Soccer data from multple countries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/miosisoniii/ussie}
  \item \url{https://miosisoniii.github.io/ussie/}
  \item Report bugs at \url{https://github.com/miosisoniii/ussie/issues}
}

}
\author{
\strong{Maintainer}: Artemio Sison \email{artemio.sison@inovio.com}

}
\keyword{internal}
