% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Work with European Football League Data}
\description{
This package allows you to work with European football league data supplied by the engsoccerdata package (Curley 2016).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NatalieWeaver/ussie}
  \item \url{https://natalieweaver.github.io/ussie/}
  \item Report bugs at \url{https://github.com/NatalieWeaver/ussie/issues}
}

}
\author{
\strong{Maintainer}: Natalie Weaver \email{natalie.weaver@outlook.com}

}
\keyword{internal}
