% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Work with European Football League Data}
\description{
This is a demo package for the "Building Tidy Tools" course at rstudio::conf(2022L).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ijlyttle/ussie}
  \item Report bugs at \url{https://github.com/ijlyttle/ussie/issues}
}

}
\author{
\strong{Maintainer}: Ian Lyttle \email{ijlyttle@me.com} (\href{https://orcid.org/0000-0001-9962-4849}{ORCID})

Authors:
\itemize{
  \item Emma Rand \email{emma.rand@york.ac.uk} (\href{https://orcid.org/0000-0002-1358-8275}{ORCID})
}

}
\keyword{internal}
