% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Working with European football data}
\description{
This is a demo package to create a sample package to work with 
          European football data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SSoleymani/ussie}
  \item Report bugs at \url{https://github.com/SSoleymani/ussie/issues}
}

}
\author{
\strong{Maintainer}: Siavash Soleymani \email{siavash.bs@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
