% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Working With European Football Data}
\description{
This is a package created during the Tidy Tools workshop at rstudio::conf(2022). Functions in this package are intended to work with data supplied by the engsoccerdata package (Curley 2016).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stevespangler/ussie}
  \item \url{https://stevespangler.github.io/ussie/}
  \item Report bugs at \url{https://github.com/stevespangler/ussie/issues}
}

}
\author{
\strong{Maintainer}: Steve Spangler \email{steven.spangler@gmail.com}

}
\keyword{internal}
