% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Work with Euro Soccer Data}
\description{
This is a demo package for the "Building Tidy Tools" workshop at rstudio::conf(2022L). It allows you to work with European soccer league data supplied by the engsoccerdata package (Curley 2016).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/supermdat/ussie}
  \item \url{https://supermdat.github.io/ussie/}
  \item Report bugs at \url{https://github.com/supermdat/ussie/issues}
}

}
\author{
\strong{Maintainer}: Daniel Turse \email{turse.mda@gmail.com}

}
\keyword{internal}
