% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: English Soccer Data Utility}
\description{
This is a demo package for the "Building Tidy Tools" workshop at at rstudio::conf(2022L). It allows you to work with European football league data supplied by the engsoccerdata package (Curley 2016).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kfkiernan/ussie}
  \item \url{https://kfkiernan.github.io/ussie/}
  \item Report bugs at \url{https://github.com/kfkiernan/ussie/issues}
}

}
\author{
\strong{Maintainer}: Kevin Kiernan \email{kkiernan73@gmail.com}

}
\keyword{internal}
