% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Work with European Football League Data}
\description{
This package allows you to work with European football league data supplied by the engsoccerdata package (Curley 2016).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/whats-in-the-box/ussie}
  \item Report bugs at \url{https://github.com/whats-in-the-box/ussie/issues}
}

}
\author{
\strong{Maintainer}: Qiu Max \email{ytqiuhaowen@gmail.com}

}
\keyword{internal}
